+++ src/consts/usage.bash
@@ -1,5 +1,5 @@
 #ignorenext
 # shellcheck shell=bash
 readonly USAGE="\
-usage: $0 ([-s] <file>)...
-   or: $0 --silent-all <file>..."
+usage: $0 [-m<cmd> [-A<arg>...]] ([-s] <file>)...
+   or: $0 [-m<cmd> [-A<arg>...]] --silent-all <file>..."