+++ CHANGELOG.md
@@ -6,6 +6,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2020-04-28 ##
+
+[1.1.0]: https://github.com/mfederczuk/utest-script/compare/v1.0.2...v1.1.0
+
+### Added ###
+
+* `-m, --command` option
+* `-A, --command-arg` option
+* `-valgrind` option
+
+### Deprecated ###
+
+* `-v` option is replaced with `-V` option
+
 ## [1.0.2] - 2019-30-01 ##
 
 [1.0.2]: https://github.com/mfederczuk/utest-script/compare/v1.0.1...v1.0.2